Step 4: permission helpers, comment responder, comments service
- check_user_in_site/channel live in user_permission; endpoints import there
- responders.comment: root_route + comment_schema_from_orm
- Materializer.comment_schema_from_orm delegates to responder
- services.comments owns create/update/upvote/delete/read
- comments endpoint is thin parse → service → return
