+++ chafan_core/app/api/api_v1/endpoints/answer_suggest_edits.py
@@ -9,7 +9,7 @@ from chafan_core.app.api import deps
 from chafan_core.app.api.api_v1.endpoints.answers import _update_answer
 from chafan_core.app.infra.request_context import RequestContext
 from chafan_core.app.common import OperationType
-from chafan_core.app.materialize import check_user_in_site
+from chafan_core.app.user_permission import check_user_in_site
 from chafan_core.app.schemas.answer import AnswerUpdate
 from chafan_core.app.schemas.answer_suggest_edit import (
     AnswerSuggestEditCreate,
