+++ chafan_core/app/api/api_v1/endpoints/channels.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Depends
 from chafan_core.app import crud, schemas
 from chafan_core.app.api import deps
 from chafan_core.app.infra.request_context import RequestContext
-from chafan_core.app.materialize import check_user_in_channel
+from chafan_core.app.user_permission import check_user_in_channel
 from chafan_core.utils.base import HTTPException_
 
 import logging
