+++ chafan_core/app/api/api_v1/endpoints/invitation_links.py
@@ -7,7 +7,7 @@ from chafan_core.app.api import deps
 from chafan_core.app.infra.request_context import RequestContext
 from chafan_core.app.common import OperationType
 from chafan_core.app.endpoint_utils import get_site
-from chafan_core.app.materialize import check_user_in_site
+from chafan_core.app.user_permission import check_user_in_site
 from chafan_core.app.schemas.invitation_link import InvitationLinkCreate
 from chafan_core.app.services import invitations as invitations_service
 from chafan_core.app.services import sites as sites_service
