+++ chafan_core/app/api/api_v1/endpoints/submission_suggestions.py
@@ -10,7 +10,7 @@ from chafan_core.app.api import deps
 from chafan_core.app.api.api_v1.endpoints.submissions import _update_submission
 from chafan_core.app.infra.request_context import RequestContext
 from chafan_core.app.common import OperationType
-from chafan_core.app.materialize import check_user_in_site
+from chafan_core.app.user_permission import check_user_in_site
 from chafan_core.app.schemas.richtext import RichText
 from chafan_core.app.schemas.submission import SubmissionUpdate
 from chafan_core.app.schemas.submission_archive import SubmissionEditableSnapshot
