+++ chafan_core/app/api/api_v1/endpoints/users.py
@@ -10,7 +10,7 @@ from chafan_core.app.infra.request_context import RequestContext
 from chafan_core.app.common import OperationType
 from chafan_core.app.endpoint_utils import get_site
 from chafan_core.app.limiter import limiter
-from chafan_core.app.materialize import check_user_in_site
+from chafan_core.app.user_permission import check_user_in_site
 from chafan_core.app.schemas.event import EventInternal, InviteJoinSiteInternal
 from chafan_core.app.services import sites as sites_service
 from chafan_core.utils.base import HTTPException_
