+++ chafan_core/app/services/__init__.py
@@ -4,6 +4,7 @@ from chafan_core.app.services import (
     answers,
     articles,
     audit,
+    comments,
     feed,
     invitations,
     link_preview,
@@ -17,6 +18,7 @@ __all__ = [
     "answers",
     "articles",
     "audit",
+    "comments",
     "feed",
     "invitations",
     "link_preview",
